requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0

# API Documentation
drf-spectacular==0.26.5

# Environment variables
python-decouple==3.8
python-dotenv==1.0.0

# Image handling
Pillow==10.1.0

# Development tools
django-extensions==3.2.3
ipython==8.17.2

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
django-filter==23.5
django-cleanup==8.0.0